requirements.txt
ipykernel
semantic-kernel
azure-search-documents
